requirements.txt
google-genai
rich
pydantic
python-dotenv
requests
chromadb
sentence-transformers
fastapi
uvicorn[standard]
streamlit
